+++ src/main/resources/static/index.html
@@ -12,8 +12,7 @@
     <script type="text/javascript" src="/webjars/bootstrap/js/bootstrap.min.js"></script>
 </head>
 <body>
-<h1>practice Social Login!!</h1>
-<h5>without spring security</h5>
+<h1>Github OAuth2 Login</h1>
 <div class="container">
     <div class="container unauthenticated">
         <br>With Github <a href="https://github.com/login/oauth/authorize?client_id=7600733c0c5ed7849ce6
